Added player.py and changed Cat class to a generic cat instead of player, npc inherits from cat
